added authorization for create/put/delete

+++ apps/api/src/country/country.controller.ts
@@ -1,5 +1,17 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common'
-import { Public } from 'nest-keycloak-connect'
+import {
+  Controller,
+  Get,
+  Post,
+  Body,
+  Patch,
+  Param,
+  Delete,
+  UnauthorizedException,
+} from '@nestjs/common'
+import { AuthenticatedUser, Public } from 'nest-keycloak-connect'
+
+import { KeycloakTokenParsed } from '../auth/keycloak'
+
 import { CountryService } from './country.service'
 import { CreateCountryDto } from './dto/create-country.dto'
 import { UpdateCountryDto } from './dto/update-country.dto'
@@ -9,8 +21,16 @@ export class CountryController {
   constructor(private readonly countryService: CountryService) {}
 
   @Post('create-country')
-  @Public()
-  create(@Body() createCountryDto: CreateCountryDto) {
+  create(
+    @AuthenticatedUser()
+    user: KeycloakTokenParsed,
+    @Body()
+    createCountryDto: CreateCountryDto,
+  ) {
+    if (!user) {
+      throw new UnauthorizedException()
+    }
+
     return this.countryService.create(createCountryDto)
   }
 
@@ -27,14 +47,31 @@ export class CountryController {
   }
 
   @Patch(':id')
-  @Public()
-  update(@Param('id') id: string, @Body() updateCountryDto: UpdateCountryDto) {
+  update(
+    @AuthenticatedUser()
+    user: KeycloakTokenParsed,
+    @Param('id')
+    id: string,
+    @Body() updateCountryDto: UpdateCountryDto,
+  ) {
+    if (!user) {
+      throw new UnauthorizedException()
+    }
+
     return this.countryService.updateCountryById(id, updateCountryDto)
   }
 
   @Delete(':id')
-  @Public()
-  remove(@Param('id') id: string) {
+  remove(
+    @AuthenticatedUser()
+    user: KeycloakTokenParsed,
+    @Param('id')
+    id: string,
+  ) {
+    if (!user) {
+      throw new UnauthorizedException()
+    }
+
     return this.countryService.removeCountryById(id)
   }
 }
